// src/app/shared/models/employee.ts
export interface Employee {
  firstName: string | null | undefined;
  lastName: string | null | undefined;
  designation: string | null | undefined;
  empId: string | null | undefined;
  gender: string | null | undefined;
  address: string | null | undefined;
  emailOffice: string | null | undefined;
  emailPersonal: string | null | undefined;
  mobileNum: string | null | undefined;
  skills: string | null | undefined[];
}